+++ traceview/app.py
@@ -3,10 +3,8 @@
 """
 traceview.app
 
-This module contains the objects associated with App Discovery
-and App Management API resources.
+This module contains the objects associated with App Management API resources.
 
-http://dev.appneta.com/docs/api-v2/discovery.html
 http://dev.appneta.com/docs/api-v2/app.html
 
 """
@@ -16,22 +14,17 @@ from .resource import Resource
 
 class Assign(Resource):
 
-    PATH = "assign_app"
+    def update(self, hostname, app, *args, **kwargs):
+        kwargs['hostname'] = hostname
+        kwargs['appname'] = app
+        return self.api.post('assign_app', *args, **kwargs)
+
 
 class App(Resource):
 
-    PATH = {
-        "GET": "apps",
-        "DELETE": "app/{app_name}"
-    }
-    def get(self, *args, **kwargs):
-        self.path = self.PATH["GET"]
-        return super(App, self).get(*args, **kwargs)
+    def get(self):
+        return self.api.get('apps')
 
     def delete(self, app_name, *args, **kwargs):
-        """ Delete an app.
-
-        :param str app_name: The name of the app to delete.
-        """
-        self.path = self.PATH["DELETE"].format(app_name=app_name)
-        return super(App, self).delete(*args, **kwargs)
+        path = 'app/{app_name}'.format(app_name=app_name)
+        return self.api.delete(path)