+++ traceview/discovery.py
@@ -14,46 +14,42 @@ from .resource import Resource
 
 class Action(Resource):
 
-    PATH = "actions"
-
-
-# for App class, see app.py
+    def get(self):
+        return self.api.get('actions')
 
 
 class Browser(Resource):
 
-    PATH = "browsers"
+    def get(self):
+        return self.api.get('browsers')
 
 
 class Controller(Resource):
 
-    PATH = "controllers"
+    def get(self):
+        return self.api.get('controllers')
 
 
 class Domain(Resource):
 
-    PATH = "domains"
+    def get(self):
+        return self.api.get('domains')
+
 
 class Layer(Resource):
 
-    PATH = "layers/{app}"
-
     def get(self, app, *args, **kwargs):
-        """ Overloaded get method.
-
-        :param str app: The app name to list layers.
-        :param int since_time: (optional) The start of the time window as a UTC timestamp in milliseconds.
-
-        """
-        self.path = self.PATH.format(app=app)
-        return super(Layer, self).get(*args, **kwargs)
+        path = 'layers/{app}'.format(app=app)
+        return self.api.get(path, *args, **kwargs)
 
 
 class Metric(Resource):
 
-    PATH = "metrics"
+    def get(self):
+        return self.api.get('metrics')
 
 
 class Region(Resource):
 
-    PATH = "regions"
+    def get(self):
+        return self.api.get('regions')